+++ packages/catalog/src/index.ts
@@ -1,6 +1,5 @@
-export * from "./i18n";
 export * from "./product";
 export * from "./attribute";
 export * from "./category";
-export * from "./queries";
-export * from "./queries";
+export * from "./attribute";
+export * from "./attribute-group";
