+++ packages/catalog/src/queries/product.types.ts
